+++ crates/module-system/sov-modules-rollup-blueprint/src/lib.rs
@@ -61,7 +61,9 @@ mod blueprint {
     use sov_rollup_interface::services::da::DaService;
     use sov_rollup_interface::storage::HierarchicalStorageManager;
     use sov_rollup_interface::zk::{ZkvmGuest, ZkvmHost};
-    use sov_sequencer::{FairBatchBuilder, Sequencer, SequencerDb, SequencerSpec};
+    use sov_sequencer::{
+        FairBatchBuilder, FairBatchBuilderConfig, Sequencer, SequencerDb, SequencerSpec,
+    };
     use sov_state::storage::NativeStorage;
     use sov_state::Storage;
     use sov_stf_runner::{
@@ -123,7 +125,11 @@ mod blueprint {
             ledger_db: &LedgerDb,
             sequencer_db: &SequencerDb,
             da_service: &Self::DaService,
-            rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+            rollup_config: &RollupConfig<
+                <Self::Spec as Spec>::Address,
+                Self::DaConfig,
+                FairBatchBuilderConfig<Self::DaSpec>,
+            >,
         ) -> Result<RuntimeEndpoints, anyhow::Error>;
 
         /// Creates GenesisConfig from genesis files.
@@ -135,7 +141,11 @@ mod blueprint {
                 Self::DaSpec,
             >>::GenesisPaths,
             kernel_genesis: <Self::Kernel as Kernel<Self::Spec, Self::DaSpec>>::GenesisConfig,
-            _rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+            _rollup_config: &RollupConfig<
+                <Self::Spec as Spec>::Address,
+                Self::DaConfig,
+                FairBatchBuilderConfig<Self::DaSpec>,
+            >,
         ) -> anyhow::Result<
             GenesisParams<
                 <Self::Runtime as RuntimeTrait<Self::Spec, Self::DaSpec>>::GenesisConfig,
@@ -156,14 +166,22 @@ mod blueprint {
         /// Creates instance of [`DaService`].
         async fn create_da_service(
             &self,
-            rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+            rollup_config: &RollupConfig<
+                <Self::Spec as Spec>::Address,
+                Self::DaConfig,
+                FairBatchBuilderConfig<Self::DaSpec>,
+            >,
         ) -> Self::DaService;
 
         /// Creates instance of [`ProverService`].
         async fn create_prover_service(
             &self,
             prover_config: RollupProverConfig,
-            rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+            rollup_config: &RollupConfig<
+                <Self::Spec as Spec>::Address,
+                Self::DaConfig,
+                FairBatchBuilderConfig<Self::DaSpec>,
+            >,
             da_service: &Self::DaService,
         ) -> Self::ProverService;
 
@@ -171,7 +189,11 @@ mod blueprint {
         /// Panics if initialization fails.
         fn create_storage_manager(
             &self,
-            rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+            rollup_config: &RollupConfig<
+                <Self::Spec as Spec>::Address,
+                Self::DaConfig,
+                FairBatchBuilderConfig<Self::DaSpec>,
+            >,
         ) -> Result<Self::StorageManager, anyhow::Error>;
 
         /// Creates instance of a LedgerDb.
@@ -193,7 +215,11 @@ mod blueprint {
                 Self::Spec ,
                 Self::DaSpec,
             >>::GenesisConfig,
-            rollup_config: RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+            rollup_config: RollupConfig<
+                <Self::Spec as Spec>::Address,
+                Self::DaConfig,
+                FairBatchBuilderConfig<Self::DaSpec>,
+            >,
             prover_config: Option<RollupProverConfig>,
         ) -> Result<Rollup<Self, M>, anyhow::Error>
         where