+++ examples/demo-rollup/src/mock_rollup.rs
@@ -18,7 +18,7 @@ use sov_risc0_adapter::host::Risc0Host;
 use sov_risc0_adapter::Risc0Verifier;
 use sov_rollup_interface::services::da::DaServiceWithRetries;
 use sov_rollup_interface::zk::aggregated_proof::CodeCommitment;
-use sov_sequencer::SequencerDb;
+use sov_sequencer::{FairBatchBuilderConfig, SequencerDb};
 use sov_state::{DefaultStorageSpec, ProverStorage, Storage, ZkStorage};
 use sov_stf_runner::{ParallelProverService, ProverService, RollupConfig, RollupProverConfig};
 use tokio::sync::watch;
@@ -80,7 +80,11 @@ impl FullNodeBlueprint<Native> for MockDemoRollup<Native> {
         ledger_db: &LedgerDb,
         sequencer_db: &SequencerDb,
         da_service: &Self::DaService,
-        rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+        rollup_config: &RollupConfig<
+            <Self::Spec as Spec>::Address,
+            Self::DaConfig,
+            FairBatchBuilderConfig<Self::DaSpec>,
+        >,
     ) -> anyhow::Result<RuntimeEndpoints> {
         let mut endpoints = sov_modules_rollup_blueprint::register_endpoints::<
             Self,
@@ -91,7 +95,7 @@ impl FullNodeBlueprint<Native> for MockDemoRollup<Native> {
             ledger_db,
             sequencer_db,
             da_service,
-            rollup_config.da.sender_address,
+            &rollup_config.sequencer,
         )?;
 
         // TODO: Add issue for Sequencer level RPC injection:
@@ -107,7 +111,11 @@ impl FullNodeBlueprint<Native> for MockDemoRollup<Native> {
 
     async fn create_da_service(
         &self,
-        rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+        rollup_config: &RollupConfig<
+            <Self::Spec as Spec>::Address,
+            Self::DaConfig,
+            FairBatchBuilderConfig<Self::DaSpec>,
+        >,
     ) -> Self::DaService {
         DaServiceWithRetries::new_fast(
             StorableMockDaService::from_config(rollup_config.da.clone()).await,
@@ -117,7 +125,11 @@ impl FullNodeBlueprint<Native> for MockDemoRollup<Native> {
     async fn create_prover_service(
         &self,
         prover_config: RollupProverConfig,
-        rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+        rollup_config: &RollupConfig<
+            <Self::Spec as Spec>::Address,
+            Self::DaConfig,
+            FairBatchBuilderConfig<Self::DaSpec>,
+        >,
         _da_service: &Self::DaService,
     ) -> Self::ProverService {
         let inner_vm = Risc0Host::new(risc0::MOCK_DA_ELF);
@@ -140,7 +152,11 @@ impl FullNodeBlueprint<Native> for MockDemoRollup<Native> {
 
     fn create_storage_manager(
         &self,
-        rollup_config: &RollupConfig<<Self::Spec as Spec>::Address, Self::DaConfig>,
+        rollup_config: &RollupConfig<
+            <Self::Spec as Spec>::Address,
+            Self::DaConfig,
+            FairBatchBuilderConfig<Self::DaSpec>,
+        >,
     ) -> anyhow::Result<Self::StorageManager> {
         NativeStorageManager::new(&rollup_config.storage.path)
     }